Checkpoint v0.8.2
Many changes to the format of the .fit_out and xls files where made in
v0.8.0 and v0.8.1.  There was a bug in v0.8.1 having to do with parsing of
date in plot_fit.py.

The problem was using try / except to first try exec of tok0=tok1 and then
using tok0 = "tok1" if that failed.  With tok0 = 'date' and tok1 =
'2016-12-15' tok0=tok1 succeded result in date=1988 (an integer).  Fix was
to use check for number (isNumber) rather than try.

This is a general problem in the parsing code.  Should be fixed.  Maybe
better would be to have lists with the type of item and do parsing based
on that.  Should also consider using dictionaries rather than using
parsing of the form exec(tok0 +'='+ tok1)